using inner class for configuration name
