+++ tests/regression/facebook/test_chrome_facebook_ail_type_comment_1_txt.py
@@ -1,12 +1,11 @@
 from lib.perfBaseTest import PerfBaseTest
-from lib.common.configName import ConfigName
 
 
 class TestSikuli(PerfBaseTest):
 
     def setUp(self):
         super(TestSikuli, self).setUp()
-        self.set_configs(ConfigName.INDEX, self.extract_platform_dep_settings({"win32": {"": {"compare-threshold": 0.003}}}))
+        self.set_configs(self.config_name.INDEX, self.extract_platform_dep_settings({"win32": {"": {"compare-threshold": 0.003}}}))
 
     def test_chrome_facebook_ail_type_comment_1_txt(self):
         self.round_status = self.sikuli.run_test(self.env.test_name, self.env.output_name, test_target=self.env.TEST_FB_HOME,
